Test fixed as per comments
