+++ src/test/java/pageobjmodel/OrderDoYouWantSubmit.java
@@ -1,7 +1,6 @@
-package pageObjModel;
+package pageobjmodel;
 
 import org.openqa.selenium.By;
-import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 
@@ -11,7 +10,6 @@ public class OrderDoYouWantSubmit {
     private final WebDriver driver;
     OrderDoYouWantSubmit(WebDriver driver){
         this.driver = driver;
-        driver.manage().timeouts().implicitlyWait(3, TimeUnit.SECONDS);
     }
 
     private static final By YES_BUTTON = By.xpath(".//button[@class='Button_Button__ra12g Button_Middle__1CSJM' and text()='Да']");
